refactor: 레이아웃과 페이지 중복에 따라 재수정
